RAG-Doc-Q&A/requirements.txt
langchain
python-dotenv
ipykernel
langchain-community
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
langchain-openai
chromadb
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
duckdb
pandas
openai
langchain-groq
duckduckgo-search
faiss-cpu
pymupdf
arxiv
wikipedia
mysql-connector-python
SQLAlchemy
validators==0.28.1
youtube_transcript_api
unstructured
pytube
numexpr
huggingface_hub
langchain_tools
cassio
PyPDF2
astrapy
transformers
gradio
boto3
awscli
langchain_nvidia_ai_endpoints
crewai
crewai_tools
yt_dlp==2023.11.14 
youtube-transcript-api==0.6.1